Implement 3 software stacks that can be used by snippets
struct CStack is a general custom stack. We use three versions of it, a
user stack, an input stack and an output stack. This ways we can push
variables to the input stack, snippets can pop those values, and push
produced values to the output stack. The user stack can be used by
snippets as they wish. The main difficulty was to implement it in a way
that does not modify the stack nor any registers.