Run in guest supervisor and user, reuse VMs
Add support for the somewhat new guest user mode.
Reuse VMs for same candidates for training and signaling. This actually
ensured that we get a singal when injecting into guest. If creating the
VM right before the signal we probably evicted something in the
microarchitecture state. But this also speeds-up the experiments a lot