Improve kmods build system
The prior setup allowed to build the kernel modules for different
kernels. I.e. host and guest at the same time. But, it rebuild them
everytime. The current system still allows to build for host and guest,
but is more tailored for building for host only. Also, the kernel
modules can be build independently and on themselves, by using a common
file.